+++ Runtime/NativeContainers/NativeExchange.cs
@@ -7,7 +7,7 @@ using Unity.Collections.LowLevel.Unsafe;
 namespace Voxell.NativeContainers
 {
   [StructLayout(LayoutKind.Sequential), NativeContainer, NativeContainerIsAtomicWriteOnly]
-  unsafe public struct NativeAdd : IDisposable
+  unsafe public struct NativeExchange : IDisposable
   {
     // the actual pointer to the allocated count needs to have restrictions relaxed
     // so jobs can be schedled with this container
@@ -29,7 +29,7 @@ namespace Voxell.NativeContainers
     // Keep track of where the memory for this was allocated
     private Allocator _allocatorLabel;
 
-    public NativeAdd(Allocator allocator)
+    public NativeExchange(Allocator allocator)
     {
       // create a dispose sentinel to track memory leaks. This also creates the AtomicSafetyHandle
       #if ENABLE_UNITY_COLLECTIONS_CHECKS
@@ -46,13 +46,13 @@ namespace Voxell.NativeContainers
       Value = 0;
     }
 
-    public void AtomicAdd(int value)
+    public int AtomicExchange(int value)
     {
       // verify that the caller has write permission on this data
       #if ENABLE_UNITY_COLLECTIONS_CHECKS
       AtomicSafetyHandle.CheckWriteAndThrow(m_Safety);
       #endif
-      Interlocked.Add(ref *_valueHolder, value);
+      return Interlocked.Exchange(ref *_valueHolder, value);
     }
 
     public int Value